feat: always show CMS "Certificate Title" on program certificates (all surfaces)
Follow-up to #3512 (merged), which gated the on-screen program certificate
title behind the cms-certificate-title flag. This makes it always-on and
extends it to the previously-ungated server-rendered surfaces:

- Removes the cms-certificate-title feature flag (no longer gated).
- The program certificate title comes from the CMS product_name ("Certificate
  Title") everywhere — on-screen cert, page heading, share text, download
  filename, the PDF route, and the page/SEO metadata — falling back to the
  program title when product_name is empty.
- Course certificates unchanged.

Part of mitodl/hq#11907.
