+++ frontends/main/src/common/feature_flags.ts
@@ -17,7 +17,6 @@ export enum FeatureFlags {
   PodcastDetailPage = "podcast-detail-page",
   B2BContractManagerDashboard = "b2b-contract-manager-dashboard",
   Arithmix = "arithmix",
-  CmsCertificateTitle = "cms-certificate-title",
 }
 
 /**
